finish: logic for Way
