+++ app.json
@@ -7,7 +7,7 @@
       "scheme": "tests",
       "web": {
         "bundler": "metro",
-        "output": "static",
+        "output": "static"
       },
       "plugins": [
       "expo-router"
@@ -38,7 +38,6 @@
     },
     "android": {
       "adaptiveIcon": {
-        "foregroundImage": "./assets/adaptive-icon.png",
         "backgroundColor": "#ffffff",
         "newArchEnabled": true
       }
