+++ database/database.json
@@ -43,7 +43,7 @@
       "phone": "+7 (900) 234-56-78",
       "tracks": [
         {
-          "date": "31 декабря",
+          "date": "3 декабря",
           "time": "10:00 - 10:10",
           "time2": "10:11 - 10:20",
           "route": [
